Move sanitize_msg_param to __init__.py
